Relocate test methods
