+++ tests/test_internals/test_glob_loader.py
@@ -1,7 +1,4 @@
 # -*- coding: utf-8 -*-
-import os
-import re
-
 import pytest
 
 from esmigrate.contexts import ContextConfig
@@ -16,11 +13,6 @@ def glob_loader():
     return _loader
 
 
-@pytest.fixture(scope='function')
-def parameter(request):
-    return request.param
-
-
 def test_glob_loader_initialized_with_test_context(glob_loader):
     assert glob_loader.get_ctx().profile == 'test'
 
@@ -32,26 +24,6 @@ def test_scan_dir_raises_not_a_directory_error(glob_loader):
 
 def test_scan_dir_parses_valid_file_names(glob_loader):
     assert len(glob_loader.scan_dir('tests/resources/schema_dir')) > 0
-
-
-@pytest.mark.parametrize('parameter', [
-    'V1_1__create_index_mapping_for_twitter.exm',
-    'V1_2__create_new_doc_in_twitter.exm',
-    'V1_3__update_existing_doc_in_twitter.exm',
-    'V1_3__update_existing_doc_in_twitter.exm',
-    'V1_4__delete_all_doc_in_twitter.exm',
-    'V1_5__delete_index_twitter.exm',
-], indirect=['parameter'])
-def test_envvar_regex(monkeypatch, parameter):
-    monkeypatch.setenv('SCHEMA_PATTERN',
-                       'V(?P<version>[\\d]+)_(?P<sequence>[\\d]+)__(?P<name>[\\w]+)\\.(?P<extension>[\\w]+)')
-
-    _context = ContextConfig()
-    _pattern = rf"^{os.getenv('SCHEMA_PATTERN')}$"
-    assert _context.schema_pattern == _pattern
-
-    rex = re.compile(_pattern)
-    assert rex.match(parameter)
 
 
 def test_scan_dir_raises_invalid_schema_pattern_error():
